requirements.txt
gradio==5.13.1
grpcio-status==1.69.0
httptools==0.6.4
langchain-chroma==0.2.1
langchain-community==0.3.15
langchain-google-genai==2.0.9
langchain-openai==0.3.1
pip==23.0.1
PySocks==1.7.1
selenium==4.27.1
setuptools==66.1.1
soupsieve==2.6
uvloop==0.21.0
watchfiles==1.0.4
youtube-transcript-api==0.6.3
